ready for eval

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: rstumpf <rstumpf@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/13 12:57:38 by rstumpf           #+#    #+#             */
-/*   Updated: 2025/01/31 16:12:37 by rstumpf          ###   ########.fr       */
+/*   Updated: 2025/01/31 18:10:03 by rstumpf          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ size_t				ft_strlen(const char *string);
 int					ft_toupper(int c);
 int					ft_tolower(int c);
 int					ft_isalnum(int c);
-long				ft_atoi(const char *str);
+int					ft_atoi(const char *str);
 char				*ft_strchr(const char *s, int c);
 char				*ft_strrchr(const char *s, int c);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
